+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Option to replace exercise in training session
+- Option to defer exercise in training session
 
 ## [0.3.0] - 2024-01-02